fish movement update

+++ TombEngine/Objects/TR3/Entity/tr3_fish_emitter.h
@@ -20,7 +20,7 @@ namespace TEN::Entities::Creatures::TR3
 		short Species;
 		bool Lethal;
 		short YAngle = ANGLE(0.0f);
-		short Timer;
+		float Timer;
 
 		Matrix Transform;
 	};
